requirements.txt
aerich==0.9.1
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
asyncclick==8.2.2.2
bcrypt==3.2.2
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dictdiffer==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
pypika-tortoise==0.6.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tortoise-orm==0.25.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
